+++ domain/src/main/java/com/sopetit/domain/entity/response/auth/LogInResponseModel.kt
@@ -1,4 +1,4 @@
-package com.sopetit.domain.entity.response
+package com.sopetit.domain.entity.response.auth
 
 data class LogInResponseModel (
     val accessToken: String = "",
